requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
docopt==0.6.2
et-xmlfile==1.1.0
ete3==3.1.3
fastapi==0.104.1
future==0.18.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.3.2
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pymongo==4.6.0
pymongo-schema==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
XlsxWriter==3.1.9
xlwt==1.3.0
gunicorn==21.2.0
